backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
requests>=2.31
httpx>=0.27

# Parsers
pdfplumber>=0.11
python-docx>=1.1
langdetect>=1.0.9
jsonschema>=4.22
PyMuPDF>=1.24

# Embeddings & NLP
sentence-transformers>=2.7
scikit-learn>=1.4
numpy>=1.26

# Pydantic v2 (FastAPI expects v2)
pydantic>=2.6
